refactor: separate board and components
Split the tkinter GUI application class 'Board' into separate classes, 'BoardTile', 'BoardMenu', 'BoardEntry', 'BoardButton', and 'BoardLabel', to improve modularity and code readability.

BoardTile: Represents a tile from the board with its logic.

BoardMenu: Represents the contextual menu in the GUI.

BoardEntry: Represents the square tile in the board.

BoardButton: Represents a solve button.

BoardLabel: Represents a result label.

These changes enhance the maintainability and extensibility of the codebase, making it more organized and readable.

Partially fixes #37
